% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{applicants}
\alias{applicants}
\title{Applicants.}
\format{A data frame with 3 variables: \code{year}, \code{sex} and
  \code{applicants}}
\usage{
applicants
}
\description{
The SSA baby names data comes from social security number (SSN) applications.
SSA cards were first issued in 1936, but were only needed for people with
an income. In 1986, the law changed effectively requiring all children to
get an SSN at birth.
}
\keyword{datasets}

