% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{babynames}
\alias{babynames}
\title{Baby names.}
\format{A data frame with five variables: \code{year}, \code{sex},
  \code{name}, \code{n} and \code{prop} (\code{n} divided by total number
  of applicants in that year).}
\usage{
babynames
}
\description{
Full baby name data provided by the SSA. This includes all names with at
least 5 uses.
}
\keyword{datasets}

