% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{births}
\alias{births}
\title{Births}
\format{A data frame:
\describe{
\item{year}{Year}
\item{births}{Number of live births, rounded to nearest 1000}}}
\usage{
births
}
\description{
Live births data from census.gov.
}
\keyword{datasets}

