% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lifetables}
\alias{lifetables}
\title{Lifetables}
\format{A data frame with nine variables: \code{x}, \code{qx}, \code{lx},
  \code{dx}, \code{Lx}, \code{Tx}, \code{ex}, \code{sex}, \code{year}}
\usage{
lifetables
}
\description{
Cohort life tables data as provided by SSA.
}
\keyword{datasets}

