% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lifetables}
\alias{lifetables}
\title{Lifetables}
\format{A data frame with nine variables:
\describe{
\item{\code{x}}{age in years}
\item{\code{qx}}{probability of death at age \code{x}}
\item{\code{lx}}{number of survivors, of birth cohort of 100,000, at next integral age}
\item{\code{dx}}{number of deaths that would occur between integral ages}
\item{\code{Lx}}{Number of person-years lived between \code{x} and \code{x+1}}
\item{\code{Tx}}{Total number of person-years lived beyond age \code{x}}
\item{\code{ex}}{Average number of years of life remaining for members of cohort alive at age \code{x}}
\item{\code{sex}}{Sex}
\item{\code{year}}{Year}
}

For further details, see \url{http://www.ssa.gov/oact/NOTES/as120/LifeTables_Body.html#wp1168594}}
\usage{
lifetables
}
\description{
Cohort life tables data as provided by SSA.
}
\keyword{datasets}

